pyproject.toml
[project]
name = "email-categorization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.79.0",
    "black>=24.0.0",
    "fastapi>=0.128.7",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "o365>=2.1.9",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.27.2",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.0",
    "pytest-json-report>=1.5.0",
    "pytest-mock>=3.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.3.0",
    "responses>=0.24.0",
    "rq>=2.7.0",
    "unicorn>=2.1.4",
    "uvicorn>=0.40.0",
]
